Fix build on MacOS

+++ source/entityhandler.d
@@ -8,10 +8,10 @@ interface EntityHandler
   void addEntity(Entity entity);
   void removeEntity(Entity entity);
   void update();
-  
+
   string debugText() @property;
   double debugTiming() @property;
-  int componentCount() @property;
-  
+  size_t componentCount() @property;
+
   string className() @property;
 }
