+++ source/eventhandlers/debuginfo.d
@@ -10,7 +10,7 @@ import systemset;
 
 void handleToggleDebugInfo(Input gameInput, SystemSet systemSet, ref Entity debugText)
 {
-  static int index = 0;
+  static size_t index = 0;
   if (gameInput.isActionToggled("toggleDebugInfo"))
   {
     if (debugText is null)
@@ -20,7 +20,7 @@ void handleToggleDebugInfo(Input gameInput, SystemSet systemSet, ref Entity debu
     }
     index = (index + 1) % systemSet.entityHandlers.length;
   }
-  
+
   // TODO: ensure entity values get reflected to the relevant components
   /*final switch (index % 3)
   {
@@ -34,8 +34,8 @@ void handleToggleDebugInfo(Input gameInput, SystemSet systemSet, ref Entity debu
       debugText.values["text"] = systemSet.graphicsTimingText;
       break;
   }*/
-  
+
   debugText.values["text"] = systemSet.entityHandlers[index].debugText;
-  
+
   assert(debugText.values["text"] !is null);
 }
